- Added skeleton PDE for perturbed fluid mechanics. - Removed some leftovers from Python/TCL scripting. - Re-added support for Pyramid and Tetra elements. Integration works up to   order 6 (while abusing order 5 as 6 for pyramids). - Default integration order is now set to 0 and mode to RELATIVE for nodal   Lagrange elements, since the order gets set to 2*(p+1) anyway. This means   that 1st order elements get integrated with order 4 and 2nd order   elements with order 6, which should account quite well, also for   distorted meshes.

+++ source/DataInOut/SimInOut/gmsh/SimOutputParsed.cc
@@ -486,7 +486,7 @@ namespace CoupledField{
       //determine next power of two available
       //cut float values
       long exp = std::ceil(log2(varsSize));
-      bfrS = std::pow(2,exp);
+      bfrS = std::pow(2.0,static_cast<Double>(exp));
     }
 
     //number of read operations
